requirements.txt
streamlit
absl-py
addict
appdirs
audioread
basicsr
cachetools
certifi
cffi
charset-normalizer
cos-python-sdk-v5
crcmod
cycler
decorator
dicttoxml
dill
facexlib
filterpy
fonttools
future
gfpgan
google-auth
google-auth-oauthlib
grpcio
idna
imageio==2.22.3
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
joblib==1.1.0
kiwisolver==1.4.4
librosa==0.9.1
llvmlite==0.38.0
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
moviepy==1.0.3
networkx==2.6.3
numba==0.55.1
numpy==1.22
oauthlib==3.2.2
opencv-contrib-python==4.5.5.62
opencv-python==4.7.0.72
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pooch==1.6.0
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
realesrgan==0.3.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SoundFile==0.10.3.post1
tb-nightly==2.11.0a20221031
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.10.0
